+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 - Support input[type=number] value data-binding as a Ruby Numeric object (Integer or Float)
 - Support input[type=range] value data-binding as a Ruby Numeric object (Integer or Float)
+- Support input[type=datetime-local] value data-binding as a Ruby Time object
 - Update Hello, Data-Binding! Sample to include a checkbox
+- New Hello, Input! Sample: `require 'glimmer-dsl-web/samples/hello/hello_input'`
 
 ## 0.0.6